fix: Mapped the return type from float to double

+++ tests/fixtures/optimizer/annotate_functions.sql
@@ -1545,7 +1545,7 @@ DOUBLE;
 
 # dialect: snowflake
 ASIN(tbl.double_col);
-FLOAT;
+DOUBLE;
 
 # dialect: snowflake
 AI_AGG('foo', 'bar');
